requirements.txt
faiss-cpu==1.8.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-text-splitters==0.0.1
regex==2023.12.25
sentence-transformers==2.6.1
openai==1.17.0
regex==2023.12.25
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.4
transformers==4.39.3
python-dotenv